pyproject.toml
[project]
name = "food-reward-analysis"
version = "0.1.0"
description = "Food reward prediction and fMRI brain activity analysis"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Yuhi Aoki" }]
dependencies = [
    # Core scientific computing
    "numpy>=2.0",
    "pandas>=2.2.3",
    "scipy>=1.13.0",
    "scikit-learn>=1.5.2",

    # Deep learning - PyTorch ecosystem
    "torch>=2.8.0",
    "torchvision>=0.19.0",
    "transformers>=4.51.3",
    "accelerate>=0.26.0",
    "peft>=0.12.0",
    "open-clip-torch>=2.32.0",

    # Visualization
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",

    # fMRI analysis
    "nilearn>=0.12.1",
    "pillow>=11.3.0",

    # Utilities
    "tqdm>=4.66.6",
    "joblib>=1.4.2",
    "dill>=0.3.9",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]
training = [
    "optuna>=4.0.0",
    "tensorboard>=2.18.0",
]
clip = [
    "ftfy>=6.3.1",
    "regex>=2024.11.6",
]

[tool.ruff]
extend-include = ["*.ipynb"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
